fix: rename course links component
